# docs/openapi.yaml
openapi: 3.1.0
info:
  title: meili-ingest Gateway API
  version: 0.1.0
  summary: Ingest any content into Meilisearch through configurable pipelines.
  description: |
    The gateway is the public entry point of meili-ingest.

    **Two tiers**

    * `POST /ingest` — auto-routing. The gateway detects the MIME type (magic
      bytes → file extension → UTF-8 sniff), selects the best matching pipeline
      (tenant-scoped user pipelines > global user pipelines > built-ins),
      resolves the target index and starts a Temporal workflow.
    * `POST /ingest/pipeline/{name}` — explicit pipeline, full control.

    **Tenant context**

    Every request is scoped to a `MeiliContext` resolved from, in order:
    `X-Meili-Host` / `X-Meili-Api-Key` / `X-Meili-Project-Id` / `X-Meili-Index`
    headers (injected by Meilisearch Cloud's Envoy and trusted only when
    `X-Meili-Envoy-Secret` matches the gateway's `ENVOY_TRUSTED_HEADER`), the
    `?index=` query parameter, an `Authorization: Bearer <key>` header, and
    finally the gateway's `MEILI_URL` / `MEILI_API_KEY` environment variables.
    When none of these yields a host and API key the request fails with `400` —
    unless the pipeline is **pinned**: every `meili_indexer` step names a
    [Meilisearch connection](/concepts/connections), which then supplies the host
    and key, and no context is needed.

    **Errors** are always `{"error": "<message>", "code": "<snake_case>"}`.
  license:
    name: MIT
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
  - {}
...
paths:
...
  /ingest/pipeline/{name}:
    post:
      tags: [Ingest]
      operationId: ingestWithPipeline
      summary: Ingest content with an explicit pipeline
      description: |
        Same body formats as `POST /ingest`, but the pipeline is chosen by `name`
        (a pipeline `uid`, e.g. `builtin.pdf` or `my-pdf-with-enrichment`).
        MIME detection still runs to fill document metadata.
      parameters:
        - name: name
          in: path
          required: true
          description: Pipeline `uid`.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9._-]+$"
          example: my-pdf-with-enrichment
        - $ref: "#/components/parameters/IndexQuery"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliHost"
        - $ref: "#/components/parameters/XMeiliApiKey"
        - $ref: "#/components/parameters/XMeiliIndex"
        - $ref: "#/components/parameters/XMeiliRegion"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MultipartIngest"
          application/json:
            schema:
              $ref: "#/components/schemas/JsonIngest"
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "202":
          $ref: "#/components/responses/IngestAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WriteUnauthorized"
        "403":
          $ref: "#/components/responses/WriteForbidden"
        "404":
          $ref: "#/components/responses/PipelineNotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
...
  /indexes/{index_uid}/ingest/pipeline/{name}:
    post:
      tags: [Ingest]
      operationId: ingestIntoIndexWithPipeline
      summary: Ingest content into a named index with an explicit pipeline
      description: Like `POST /ingest/pipeline/{name}`; the path's index wins over the pipeline trigger's `index_pattern`.
      parameters:
        - name: name
          in: path
          required: true
          description: Pipeline `uid`.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9._-]+$"
        - $ref: "#/components/parameters/IndexUid"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliHost"
        - $ref: "#/components/parameters/XMeiliApiKey"
        - $ref: "#/components/parameters/XMeiliRegion"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MultipartIngest"
          application/json:
            schema:
              $ref: "#/components/schemas/JsonIngest"
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "202":
          $ref: "#/components/responses/IngestAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WriteUnauthorized"
        "403":
          $ref: "#/components/responses/WriteForbidden"
        "404":
          $ref: "#/components/responses/PipelineNotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
...
  /pipelines:
    get:
      tags: [Pipelines]
      operationId: listPipelines
      summary: List pipelines
      description: Built-in pipelines, global user pipelines and — when a project id is present — that tenant's pipelines.
      parameters:
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "200":
          description: All visible pipelines.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineDefinition"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
    post:
      tags: [Pipelines]
      operationId: upsertPipeline
      summary: Create or update a pipeline
      description: |
        Accepts YAML (`application/x-yaml`, `text/yaml`) or JSON. The definition
        is normalised (implicit sequential `depends_on`), validated (unknown
        plugin, cycle, bad `fan_out` → `422`) and upserted by `(uid, tenant_id)`;
        `version` is bumped on update. `uid`s starting with `builtin.` are
        rejected with `403`. Behind Envoy the pipeline is scoped to the calling
        project.
      parameters:
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineDefinition"
          application/x-yaml:
            schema:
              type: string
            example: |
              uid: my-pdf-with-enrichment
              name: "PDF with LLM enrichment"
              trigger:
                content_types: [application/pdf]
                filename_pattern: "contract_*.pdf"
              steps:
                - id: extract
                  plugin: pdf_extractor
                  config: { per_page: true }
                - id: chunk
                  plugin: chunker
                  depends_on: [extract]
                - id: enrich
                  plugin: llm_enricher
                  depends_on: [chunk]
                  fan_out: "$.documents"
                  config: { model: gpt-4o-mini }
                - id: index
                  plugin: meili_indexer
                  depends_on: [enrich]
          text/yaml:
            schema:
              type: string
      responses:
        "201":
          description: The stored (normalised) definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineDefinition"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Attempt to write a `builtin.*` pipeline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { "error": "builtin pipelines are read-only", "code": "forbidden" }
        "422":
          description: Invalid pipeline (unknown plugin, cycle, duplicate step, unsupported `fan_out`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { "error": "step \"enrich\" references unknown plugin \"llm_enrichr\"", "code": "invalid_pipeline" }
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
  /pipelines/{uid}:
    get:
      tags: [Pipelines]
      operationId: getPipeline
      summary: Get a pipeline
      parameters:
        - $ref: "#/components/parameters/PipelineUid"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "200":
          description: The pipeline definition (a tenant pipeline shadows a global one with the same uid).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineDefinition"
        "404":
          $ref: "#/components/responses/PipelineNotFound"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
    delete:
      tags: [Pipelines]
      operationId: deletePipeline
      summary: Delete a user pipeline
      description: |
        Never blocked by sources that use the pipeline: they are archived (their
        schedules removed) and remain readable with `archived_at` set.
      parameters:
        - $ref: "#/components/parameters/PipelineUid"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "204":
          description: Deleted.
        "403":
          description: Built-in pipelines cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { "error": "builtin pipelines cannot be deleted", "code": "forbidden" }
        "404":
          $ref: "#/components/responses/PipelineNotFound"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Self-hosted fallback: a Meilisearch API key with write access. Used as
        `api_key` when no `X-Meili-Api-Key` header is present. Combine with
        `X-Meili-Host` (or the gateway's `MEILI_URL`) to target an instance.
  parameters:
    IndexUid:
      name: index_uid
      in: path
      required: true
      description: Target index. Wins over every other index source. Meilisearch uid rules apply (alphanumeric, `-`, `_`, at most 400 bytes).
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]{1,400}$"
      example: movies
    IndexQuery:
      name: index
      in: query
      required: false
      description: Target index. Overrides `X-Meili-Index`; overridden by the pipeline trigger's `index_pattern`.
      schema:
        type: string
      example: contracts
...
    PipelineUid:
      name: uid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9._-]+$"
      example: my-pdf-with-enrichment
...
    XMeiliProjectId:
      name: X-Meili-Project-Id
      in: header
      required: false
      description: Tenant / project identifier (Envoy-injected).
      schema:
        type: string
      example: xxx
    XMeiliHost:
      name: X-Meili-Host
      in: header
      required: false
      description: Full Meilisearch URL for this tenant (Envoy-injected, or set manually in standalone mode).
      schema:
        type: string
        format: uri
      example: https://xxx.us-west.meilisearch.io
    XMeiliApiKey:
      name: X-Meili-Api-Key
      in: header
      required: false
      description: Meilisearch API key with write access (Envoy-injected).
      schema:
        type: string
    XMeiliIndex:
      name: X-Meili-Index
      in: header
      required: false
      description: Starting point of the index resolution chain (Envoy-injected, from path or query).
      schema:
        type: string
      example: documents
    XMeiliRegion:
      name: X-Meili-Region
      in: header
      required: false
      description: Region tag for observability (Envoy-injected).
      schema:
        type: string
      example: us-west
    XMeiliEnvoySecret:
      name: X-Meili-Envoy-Secret
      in: header
      required: false
      description: |
        Shared secret proving the request came through Envoy. Required for the
        other `X-Meili-*` headers to be honoured when the gateway runs with
        `ENVOY_TRUSTED_HEADER` set; ignored otherwise.
      schema:
        type: string
  responses:
    IngestAccepted:
      description: Job accepted and queued.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IngestResponse"
          example:
            job_id: 550e8400-e29b-41d4-a716-446655440000
            pipeline_used: builtin.pdf
            target_index: documents
            status: queued
    BadRequest:
      description: Malformed request or missing tenant context (no `X-Meili-*` headers, no bearer key and no `MEILI_URL`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "no Meilisearch context: set X-Meili-Host/X-Meili-Api-Key headers or MEILI_URL/MEILI_API_KEY", "code": "missing_context" }
    WriteUnauthorized:
      description: "`WRITE_PREFLIGHT` is on and the request carries no Meilisearch key Meilisearch accepts (it answered 401)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "the API key cannot add documents to index \"movies\": The Authorization header is missing. It must use the bearer authorization method.", "code": "unauthorized" }
    WriteForbidden:
      description: "`WRITE_PREFLIGHT` is on and the request's Meilisearch key may not add documents to, or read, the target index. No job was queued."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "the API key cannot add documents to index \"hn\": The API key cannot acces the index `hn`, authorized indexes are [\"glutony-*\"].", "code": "forbidden" }
    PayloadTooLarge:
      description: Body exceeds `MAX_UPLOAD_MB`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "upload exceeds 500 MiB", "code": "payload_too_large" }
...
    PipelineNotFound:
      description: Unknown pipeline.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "pipeline \"nope\" not found", "code": "not_found" }
...
    Unprocessable:
      description: Validation failed (bad cron or template, forbidden host — including an ingest `url` outside `SOURCE_FETCH_HOSTS` — unreachable Meilisearch or rejected key, unpinned pipeline).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "pipeline \"movies\" cannot run on a schedule: its meili_indexer step names no Meilisearch `connection`", "code": "validation" }
...
    UpstreamUnavailable:
      description: Control plane or Temporal unreachable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "control plane unreachable", "code": "upstream_error" }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message.
        code:
          type: string
          description: Stable snake_case error code.
          examples: [missing_context, bad_request, forbidden, not_found, payload_too_large, unsupported_media_type, invalid_pipeline, upstream_error, internal_error]
...
    MultipartIngest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: The content to ingest. Filename and `Content-Type` of the part are used as MIME hints.
        index:
          type: string
          description: Target index (same as `?index=`).
        filename:
          type: string
          description: Overrides the part's filename.
    JsonIngest:
      type: object
      description: Exactly one of `url`, `s3` or `documents` must be present.
      properties:
        url:
          type: string
          format: uri
          description: |
            HTTP(S) URL fetched by the worker. Must satisfy `SOURCE_FETCH_HOSTS` (default
            `public`: `https`, public addresses only), checked here — a forbidden URL is a
            `422` — and again at fetch time on every redirect hop.
        s3:
          type: string
          description: "A cloud object-store URI: `s3://`, `s3a://`, `gs://`, `az://`, `adl://`, `azure://`, `abfs://` or `abfss://`, with a bucket. Any other scheme (`file://`, `memory://`, …) is a `422`. Streamed by the worker with its ambient cloud credentials: not subject to `SOURCE_FETCH_HOSTS`, so scope those credentials to what callers may read."
        documents:
          type: array
          description: Inline documents, indexed directly (no extraction). Must be JSON objects with an `id`.
          items:
            type: object
            additionalProperties: true
        filename:
          type: string
          description: Filename hint used for MIME detection and `_meta.filename`.
        content_type:
          type: string
          description: MIME hint, used only when magic bytes and extension are inconclusive.
        index:
          type: string
          description: Target index (same as `?index=`).
      oneOf:
        - required: [url]
        - required: [s3]
        - required: [documents]
...
    IngestResponse:
      type: object
      required: [job_id, pipeline_used, target_index, status]
      properties:
        job_id:
          type: string
          format: uuid
        pipeline_used:
          type: string
          description: "`uid` of the pipeline that will run."
          example: builtin.pdf
        target_index:
          type: string
          description: Fully resolved Meilisearch index name.
          example: documents
        status:
          $ref: "#/components/schemas/JobStatus"
...
    PipelineDefinition:
      type: object
      required: [uid, steps]
      properties:
        uid:
          type: string
          pattern: "^[a-zA-Z0-9._-]+$"
          description: Unique id. `builtin.*` is reserved.
          example: my-pdf-with-enrichment
        name:
          type: string
          default: ""
          description: Display name (defaults to `uid`).
        description:
          type: string
        version:
          type: integer
          minimum: 1
          default: 1
          description: Bumped by the control plane on every update.
        trigger:
          $ref: "#/components/schemas/PipelineTrigger"
        steps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StepDefinition"
        builtin:
          type: boolean
          default: false
          readOnly: true
        tenant_id:
          type: string
          description: Tenant scope. Absent = global.
          readOnly: true
...